Fixed treatment of exit portals. Portals now include the neighbourhood of the exit site (i.e. 27 sites in all). Portal positions are regularly adjusted to ensure: (1) that a portal is close to the blob boundary, and all 27 sites are inside (2) that portals do not get too close to each other.
The calibration has been redone for the chemotaxis with chemo_K_exit=0 case, Tres=24.
The prob of a cell exitting (exit_prob) was reduced to 0.02, to increase the number
of exit portals and reduce the effect of an incremental change to Nexits.
The dependence of Fe on Ncells was established, and quadratic fitting gave the coeffs.
a, b, c used to determine Fe(Ncells).  Adjustment for Tres is included.

The treatment of vascularity was corrected (it was wrong, could settle to steady-state with V /= 1).
This involved changing most of the vascularity parameters.  Also added a lower bound of 1.0 on V.
It isn't clear that this is still needed.

These parameter values ensure steady-state with no inflammation, and return to steady-state
when inflammation dies away.